// apps/api/src/db/schema.ts
export type Numeric = ColumnType<string, number | string, number | string>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;
// ...
export interface VGoalStatus {
  created_at: Timestamp | null;
  current_funded_amount: Numeric | null;
  customer_id: string | null;
  funded_pct: Numeric | null;
  goal_id: string | null;
  goal_name: string | null;
  goal_type: string | null;
  is_high_priority_underfunded: boolean | null;
  is_name_type_mismatch: boolean | null;
  is_overdue: boolean | null;
  is_overfunded: boolean | null;
  priority: string | null;
  target_amount: Numeric | null;
  target_date: string | null;
  updated_at: Timestamp | null;
}